Design MainPage

+++ ForecastApp/ForecastApp/Models/WeatherData.cs
@@ -13,7 +13,7 @@ namespace ForecastApp.Models
         public double sea_level { get; set; } 
         public double grnd_level { get; set; } 
         public int humidity { get; set; } 
-        public int temp_kf { get; set; } 
+        public double temp_kf { get; set; } 
 
     }
